+++ app/favorites/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Favorites() {
+  return (
+    <div>Favorites</div>
+  )
+}
