+++ app/promos/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Promo() {
+  return (
+    <div>Promos</div>
+  )
+}
