+++ app/shop/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Shop() {
+  return (
+    <div>Shop</div>
+  )
+}
