--- components/tabs/discover.tsx
@@ -1,9 +1,0 @@
-export default function Discover() {
-    return (
-        <>
-            <div className="bg-blue-300">
-                Discover    
-            </div>        
-        </>
-    )
-}
