--- components/tabs/promo.tsx
@@ -1,9 +1,0 @@
-export default function Promos() {
-    return (
-        <>
-            <div className="bg-red-300">
-                Promos    
-            </div>        
-        </>
-    )
-}
