--- components/tabs/shop.jsx
@@ -1,9 +1,0 @@
-export default function Shop() {
-    return (
-        <>
-            <div className="">
-                Shop    
-            </div>        
-        </>
-    )
-}
